Arrastar e Selecionar Imagens

+++ src/EventListener.cpp
@@ -92,7 +92,7 @@ void EventListener::mouse_left(int state) {
     }
 }
 
-void EventListener::mouse_move(Vector2<int> pos, Vector2<int> disp) {
+void EventListener::mouse_move(Vector2<float> pos, Vector2<float> disp) {
     for (const  auto item: EventListener::EventMatrix[en_mouse_move]){
         bool stop = item.second->mouse_move(pos, disp);
